+++ src/main/resources/application.properties
@@ -10,3 +10,5 @@ spring.datasource.password=1111
 spring.jpa.database=mysql
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=create-drop
+
+spring.jwt.secret=secretKey
